univeristy_gofundme/src/uni_gofundme/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.2.1
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
